requirements.txt
Flask
googletrans==4.0.0-rc1
torch
diffusers
transformers
Pillow
opencv-python
